refactor(user): user 관련 로직을 auth에서 user domain으로 이동
